+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 - [#820](https://github.com/hyperf/hyperf/pull/820) Added nats component.
 - [#832](https://github.com/hyperf/hyperf/pull/832) Added `Hyperf\Utils\Codec\Json`.
 - [#833](https://github.com/hyperf/hyperf/pull/833) Added `Hyperf\Utils\Backoff`.
-- [#852](https://github.com/hyperf/hyperf/pull/852) Added a `clear()` method for `Hyperf\Utils\Parallel` to clear adde callbacks.
+- [#852](https://github.com/hyperf/hyperf/pull/852) Added a `clear()` method for `Hyperf\Utils\Parallel` to clear added callbacks.
 - [#854](https://github.com/hyperf/hyperf/pull/854) Added `GraphQLMiddleware`.
+- [#873](https://github.com/hyperf/hyperf/pull/873) Added redis cluster.
 
 ## Fixed